Fix issue where expt followed by division could result in error.
We should not assume that the numbers in the stack at any given
time can be cast directly to a BigDecimal, because some operations,
(like expt) may return a double or something, and then suddenly
one of the operands is not castable.

Instead, in the division function, explicitly create BigDecimals
from the operands and work with those.